pyproject.toml
[project]
name = "streamlit-mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.51.0",
    "fastapi>=0.115.12",
    "fastmcp>=2.3.4",
    "mcp>=1.9.0",
    "openai>=1.79.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
